Updated CI to publish nuget package
Switched back to windows-latest because ubuntu doesn't have nuget installed...

As a result, needed to switch back to 7z and hub

Pushes packages to the nuget package repo of the owner of the repo
running this action. Projects do not have package repos, but github uses
a neat filter to display the packages built for the project.